Moved some Startup.cs logic into new classes.
