+++ src/API/Loggy.Api/Startup.cs
@@ -1,6 +1,7 @@
 ﻿using GraphiQl;
 using GraphQL;
 using GraphQL.Types;
+using Loggy.Api.Config;
 using Loggy.Api.DataAccess;
 using Loggy.Api.Model;
 using Loggy.Api.Model.Queries;
@@ -25,8 +26,12 @@ namespace Loggy.Api
 		public void ConfigureServices(IServiceCollection services)
 		{
 			services.AddMvc();
-			RegisterRepositoriesAndDataAccess(services);
-			RegisterGraphQlTypes(services);
+			
+			new RepoAndDataAccessDependencyInjectionConfig()
+				.RegisterForDependencyInjection(services, Configuration);
+			
+			new GraphQlDependencyInjectionConfig()
+				.RegisterForDependencyInjection(services);
 		}
 
 		public void Configure(IApplicationBuilder app, IHostingEnvironment env)
@@ -35,52 +40,6 @@ namespace Loggy.Api
 
 			app.UseGraphiQl();
 			app.UseMvc();
-		}
-
-		private static void RegisterGraphQlTypes(IServiceCollection services)
-		{
-			services.AddSingleton<IDependencyResolver>(s => new FuncDependencyResolver(s.GetRequiredService));
-			services.AddSingleton<IDocumentExecuter, DocumentExecuter>();
-
-			RegisterQueryServices(services);
-			RegisterGraphTypes(services);
-
-			var serviceProvider = services.BuildServiceProvider();
-			services.AddSingleton<ISchema>(
-				new Schema(new FuncDependencyResolver(type => serviceProvider.GetService(type))));
-		}
-
-		private void RegisterRepositoriesAndDataAccess(IServiceCollection services)
-		{
-			services.AddTransient<IMongoClient, MongoClient>(sp => new MongoClient(Configuration["DbConnectionString"]));
-
-			services.AddTransient<ILogEntryRepository, LogEntryRepository>();
-			services.AddTransient<IUserRepository, UserRepository>();
-			services.AddTransient<ILogSubjectRepository, LogSubjectRepository>();
-			services.AddTransient<IDataTypeRepository, DataTypeRepository>();
-		}
-
-		private static void RegisterQueryServices(IServiceCollection services)
-		{
-			services.AddSingleton<RootQuery>();
-			services.AddSingleton<LogEntriesQuery>();
-			services.AddSingleton<LogSubjectsQuery>();
-			services.AddSingleton<UsersQuery>();
-		}
-
-		private static void RegisterGraphTypes(IServiceCollection services)
-		{
-			//Log Entries are related to Users:
-			services.AddSingleton<UserGraphType>();
-
-			//Log Entries have one or more LogFieldEntries:
-			services.AddSingleton<LogEntryGraphType>();
-			services.AddSingleton<LogFieldEntryGraphType>();
-
-			//Log Entries represent LogSubjectDefinitions with Fields (which have Data Types).
-			services.AddSingleton<LogSubjectDefinitionGraphType>();
-			services.AddSingleton<LogFieldDefinitionGraphType>();
-			services.AddSingleton<DataTypeDefinitionGraphType>();
 		}
 	}
 }
